+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -418,7 +418,7 @@ class CreateDatasourceResponseBody(BaseModel):
     metadata: CreateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYSCD1TSBY4JRR3Y8R8VCCFB"
+        "01KYV6VQ2NJ3YKBSHJ17AT0MGH"
     )
     r"""The unique identifier of the data source"""