+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsejsonschema.py
@@ -40,9 +40,9 @@ class EvaluatorResponseJSONSchema(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-30T11:26:12.017Z"
+    created: Optional[str] = "2026-07-31T04:27:48.510Z"
 
-    updated: Optional[str] = "2026-07-30T11:26:12.017Z"
+    updated: Optional[str] = "2026-07-31T04:27:48.510Z"
 
     updated_by_id: OptionalNullable[str] = UNSET