+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponseragas.py
@@ -59,9 +59,9 @@ class EvaluatorResponseRagas(BaseModel):
 
     model: str
 
-    created: Optional[str] = "2026-07-30T11:26:12.017Z"
+    created: Optional[str] = "2026-07-31T04:27:48.510Z"
 
-    updated: Optional[str] = "2026-07-30T11:26:12.017Z"
+    updated: Optional[str] = "2026-07-31T04:27:48.510Z"
 
     updated_by_id: OptionalNullable[str] = UNSET