+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -247,7 +247,7 @@ class ListDatasourcesData(BaseModel):
     metadata: ListDatasourcesMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYSCD1TP48HJ052WP6AH7J2R"
+        "01KYV6VQ2J2VY1BA911T2PC107"
     )
     r"""The unique identifier of the data source"""