+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -169,7 +169,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     metadata: RetrieveDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYSCD1TR1V06HM4G00K9CYA8"
+        "01KYV6VQ2MS2DP1M6ER4HW5C14"
     )
     r"""The unique identifier of the data source"""