+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -183,7 +183,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     metadata: UpdateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYSCD1TVK182DE5DSPVACFJH"
+        "01KYV6VQ2R4S9JT04BHB5KDZB7"
     )
     r"""The unique identifier of the data source"""